+++ src/main/java/compedia/vn/tickmi/mail/entity/Ticket.java
@@ -15,16 +15,16 @@ public class Ticket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "TICKET_ID")
-    private Integer ticketId;
+    private Long ticketId;
 
     @Column(name = "PROVIDER_ID")
     private Integer providerId;
 
     @Column(name = "TICKET_EVENT_ID")
-    private Integer ticketEventId;
+    private Long ticketEventId;
 
     @Column(name = "EVENT_ID")
-    private Integer eventId;
+    private Long eventId;
 
     @Column(name = "PATH_QR")
     private String pathQr;
@@ -45,7 +45,7 @@ public class Ticket {
     private Integer status;
 
     @Column(name = "OBJECT_ID")
-    private Integer objectId;
+    private Long objectId;
 
     @Column(name = "TYPE")
     private Integer type;
