+++ src/main/java/compedia/vn/tickmi/mail/task/qr/GenerateQR.java
@@ -31,7 +31,7 @@ public class GenerateQR {
 
 
 
-    public static String handlerGeneratePathQR(String ticketEventCode, Integer eventId, String nameTicket) {
+    public static String handlerGeneratePathQR(String ticketEventCode, Long eventId, String nameTicket) {
         String pathQR = null;
         String root = PropertiesUtil.getProperty("vn.cpa.static.location.upload.gen_qr");
         String filePathOutPut = PropertiesUtil.getProperty("vn.cpa.static.location.export.qr");
